+++ warden/app/vemanager/vemanager.go
@@ -12,12 +12,11 @@ import (
 )
 
 // VoteExtensionManager wraps multiple ABCI handlers (ExtendVoteHandler,
-// VerifyVoteExtensionHandler, PrepareProposalHandler) and exposes a single
+// VerifyVoteExtensionHandler) and exposes a single
 // ABCI handler that aggregates all of them.
 type VoteExtensionManager struct {
 	extendVoteHandler          []sdk.ExtendVoteHandler
 	verifyVoteExtensionHandler []sdk.VerifyVoteExtensionHandler
-	prepareProposalHandler     []sdk.PrepareProposalHandler
 }
 
 // NewVoteExtensionManager creates a new empty VoteExtensionManager.
@@ -32,11 +31,9 @@ func NewVoteExtensionManager() *VoteExtensionManager {
 func (m *VoteExtensionManager) Register(
 	extendVoteHandler sdk.ExtendVoteHandler,
 	verifyVoteExtensionHandler sdk.VerifyVoteExtensionHandler,
-	prepareProposalHandler sdk.PrepareProposalHandler,
 ) {
 	m.extendVoteHandler = append(m.extendVoteHandler, extendVoteHandler)
 	m.verifyVoteExtensionHandler = append(m.verifyVoteExtensionHandler, verifyVoteExtensionHandler)
-	m.prepareProposalHandler = append(m.prepareProposalHandler, prepareProposalHandler)
 }
 
 // ExtendVoteHandler returns the ExtendVoteHandler that aggregates all the
@@ -100,31 +97,6 @@ func (m *VoteExtensionManager) VerifyVoteExtensionHandler() sdk.VerifyVoteExtens
 		}
 
 		return combineResponseVerifyVoteExtension(resps), nil
-	}
-}
-
-// PrepareProposalHandler returns the PrepareProposalHandler that aggregates
-// all the registered PrepareProposalHandlers.
-func (m *VoteExtensionManager) PrepareProposalHandler() sdk.PrepareProposalHandler {
-	return func(ctx sdk.Context, req *cometabci.RequestPrepareProposal) (*cometabci.ResponsePrepareProposal, error) {
-		var resp *cometabci.ResponsePrepareProposal
-		for _, handler := range m.prepareProposalHandler {
-			if resp != nil {
-				// handlers are a pipeline, so the txs returned by the previous
-				// handler are the txs that will be included in the request for
-				// the next
-				req.Txs = resp.Txs
-			}
-
-			scopedResp, err := handler(ctx, req)
-			if err != nil {
-				return nil, err
-			}
-
-			resp = scopedResp
-		}
-
-		return resp, nil
 	}
 }
